Start Time and End Time is now linked with the Schema and will update the schema accordingly
issue/210 - Extend formatting toolchain to work with both projects (#217)

* Remove husky from doto-frontend

* Initialise multi package project with lerna

* Simplify node modules installation with one command

* Configure husky on root package

* Configure doto-frontend package for precommit lint-staged

* Configure doto-backend package for precommit lint-staged

* Configure eslintrc for mocha testing

* Format all doto-backend files

* Reformat files after rebase

* Update readme with lerna and yarn instructions

Forgot to change variable names in DotoServices and fixed conflict issue in User.js

Created a new ActiveHoursContext. Passed the context to the settings and calender page. State for the times are maintained in Route.js file instead of SettingsPage.js file

Fixed import for ActiveHoursContext. Changed PropType for WorkingHoursPicker.

Updated user scheme to change type of startTime and endTime fields. Added the context consumers for ActiveHoursContext in Calendar.js. Added parameters for the startTime and endTime for the method in TaskScheduler.

Added startTime and endTime to taskscheduler can use it

shift scheduled tasks based on working hours

Fixed Magic numbers and some variable names to fix the camelcase format

changed numbers and  string constants to become all capitals